NotebookAI/backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
databases[postgresql]==0.8.0

# Redis & Caching
redis==5.0.1
celery==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0

# AI & ML
openai==1.3.7
sentence-transformers==2.2.2
qdrant-client==1.6.9
langchain==0.0.340
tiktoken==0.5.2

# Media Processing
ffmpeg-python==0.2.0
whisper==1.1.10

# Object Storage
minio==7.2.0
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1